pyproject.toml
[tool.poetry]
name = "mentalmaths"
version = "0.1.0"
description = ""
authors = ["Victor Barna <victorbarna@hotmail.fr>"]
readme = "README.md"

packages = [{ include = "mental_maths", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26"
matplotlib = "^3.8"
streamlit = "^1.43.1"
pyyaml = "^6.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
